Version 2
Images are working!

The GUI if officially coded for looks. The GUI is coded to fit as closely to the Figma mocks as possible. This was done using my personal phone.

Adjustment arrows need to be coded correctly to allow for user altering the "desiredTemp" for the GUI to be fully working pre backend.